Moved code, static files and templates into a Python package.
