+++ cicero/images.py
