+++ cicero/static/fonts/glyphicons-halflings-regular.eot
