+++ cicero/static/js/github.js
